refactor github code into a separate concern to aid testing
